Adding lpcc, an LPC compiler/disassembler
